Update autoreload to fire on more fsnotify events
When executed within a docker container, we sometimes only receive the
CHMOD event, which was being ignored. This change modifies the
autoreload to automatically reload on any event. However CHMOD events
can trigger from running the command, so a delay needed to be added as
well to avoid triggering a reload in a loop.